Separate files for base58string classes
